# perf: wallet balance in dash-qt