--- a/src/wallet/interfaces.cpp
+++ b/src/wallet/interfaces.cpp
@@ -432,7 +432,7 @@ public:
     CAmount getAvailableBalance(const CCoinControl& coin_control) override
     {
         if (coin_control.IsUsingCoinJoin()) {
-            return m_wallet->GetBalance(0, coin_control.m_avoid_address_reuse, false, &coin_control).m_anonymized;
+            return m_wallet->GetBalanceAnonymized(coin_control);
         } else {
             return m_wallet->GetAvailableBalance(&coin_control);
         }