--- a/src/wallet/transaction.h
+++ b/src/wallet/transaction.h
@@ -291,6 +291,7 @@ public:
 
     struct BalanceAnonymized
     {
+        CAmount m_anonymized{0};
         CAmount m_denom_credit{0};
         bool is_unconfirmed{false};
     };
@@ -299,9 +300,8 @@ public:
     // annotation "EXCLUSIVE_LOCKS_REQUIRED(pwallet->cs_wallet)". The
     // annotation "NO_THREAD_SAFETY_ANALYSIS" was temporarily added to avoid
     // having to resolve the issue of member access into incomplete type CWallet.
-    CAmount GetAnonymizedCredit() const NO_THREAD_SAFETY_ANALYSIS;
     CAmount GetAnonymizedCredit(const CCoinControl& coinControl) const NO_THREAD_SAFETY_ANALYSIS;
-    BalanceAnonymized GetDenominatedCredit() const NO_THREAD_SAFETY_ANALYSIS;
+    BalanceAnonymized GetAnonymizedBalance() const NO_THREAD_SAFETY_ANALYSIS;
 
     /** Get the marginal bytes if spending the specified output from this transaction */
     int GetSpendSize(unsigned int out, bool use_max_sig = false) const