## Issue being fixed or feature implemented
Wallets with big amount of transactions are loading too long time.

## What was done?
Optimized heavy calculations of balance by Wallet. See commits for details.
It makes Dash-Qt also a bit more responsible because a lot of balance calculations are done in `main` thread to update balances, though "the lagging" is not easy to measure.

## How Has This Been Tested?
Import pubkey with a lot transaction, for example `yVXDAM73Tg6A44Bm3qduXsMCYxzuqBCT48`
call `unloadwallet` + `loadwallet`, check time.

Loading of wallet is speed up from 24seconds to 23.5 seconds (average by mutiple runs).

here's detailed breakdown to particular functions (see branch with extra logs https://github.com/knst/dash/commits/perf-wallet-balance-with-logs/):
```
origin/develop (2f15c2bd81c65bba399b03a8c5186540412d7c7d)
2025-02-01T11:12:29Z [qt-rpcconsole] [w1] Wallet completed loading in            3279ms
2025-02-01T11:12:29Z [qt-rpcconsole] [w1] Rescan completed in              43ms
2025-02-01T11:12:43Z [      main] knst getBalances() 589.92ms
2025-02-01T11:12:43Z [      main] knst getAnonymizableBalance() 173.80ms
2025-02-01T11:12:43Z [      main] knst getDenominatedBalance() 186.28ms
2025-02-01T11:12:43Z [      main] knst getDenominatedBalance() 186.01ms
2025-02-01T11:12:43Z [      main] knst getNormalizedAnonymizedBalance() 27.66ms
2025-02-01T11:12:43Z [      main] knst getAverageAnonymizedRounds() 37.94ms
2025-02-01T11:12:44Z [      main] knst getAnonymizableBalance() 174.10ms
2025-02-01T11:12:44Z [      main] knst getDenominatedBalance() 189.23ms
2025-02-01T11:12:44Z [      main] knst getDenominatedBalance() 187.73ms
2025-02-01T11:12:44Z [      main] knst getNormalizedAnonymizedBalance() 27.77ms
2025-02-01T11:12:44Z [      main] knst getAverageAnonymizedRounds() 38.13ms
2025-02-01T11:12:47Z [      main] knst getBalances() 187.53ms
2025-02-01T11:12:47Z [      main] knst getBalances() 195.72ms
2025-02-01T11:12:48Z [      main] knst getBalances() 192.20ms
2025-02-01T11:12:48Z [      main] knst getBalances() 184.09ms
2025-02-01T11:12:48Z [      main] knst getAnonymizableBalance() 171.66ms
2025-02-01T11:12:48Z [      main] knst getDenominatedBalance() 185.26ms
2025-02-01T11:12:48Z [      main] knst getDenominatedBalance() 186.41ms
2025-02-01T11:12:48Z [      main] knst getNormalizedAnonymizedBalance() 25.77ms
2025-02-01T11:12:48Z [      main] knst getAverageAnonymizedRounds() 36.41ms
2025-02-01T11:12:49Z [      main] knst getBalances() 190.71ms
2025-02-01T11:12:49Z [      main] knst getAnonymizableBalance() 171.36ms
2025-02-01T11:12:49Z [      main] knst getDenominatedBalance() 185.21ms
2025-02-01T11:12:49Z [      main] knst getDenominatedBalance() 185.09ms
2025-02-01T11:12:49Z [      main] knst getNormalizedAnonymizedBalance() 26.28ms
2025-02-01T11:12:49Z [      main] knst getAverageAnonymizedRounds() 36.08ms

In total: 3279  +43  +589.92  +173.80  +186.28  +186.01  +27.66  +37.94  +174.10  +189.23  +187.73  +27.77  +38.13  +187.53  +195.72  +192.20  +184.09  +171.66  +185.26  +186.41  +25.77  +36.41  +190.71  +171.36  +185.21  +185.09  +26.28  +36.08
=7300.35
```

This PR:
```
perf-wallet-balance  (b0f71b4d3c714671a7174f7521988030a7a5708f)
# new HEAD
2025-02-01T11:20:11Z [qt-rpcconsole] [w1] Wallet completed loading in            3404ms
2025-02-01T11:20:11Z [qt-rpcconsole] [w1] Rescan completed in               0ms
2025-02-01T11:20:24Z [      main] knst getBalances() 544.79ms
2025-02-01T11:20:25Z [      main] knst getAnonymizableBalance() 173.77ms
2025-02-01T11:20:25Z [      main] knst getBalances() 139.59ms
2025-02-01T11:20:25Z [      main] knst getNormalizedAnonymizedBalance() 25.87ms
2025-02-01T11:20:25Z [      main] knst getAverageAnonymizedRounds() 35.35ms
2025-02-01T11:20:25Z [      main] knst getAnonymizableBalance() 171.82ms
2025-02-01T11:20:25Z [      main] knst getBalances() 141.07ms
2025-02-01T11:20:25Z [      main] knst getNormalizedAnonymizedBalance() 25.77ms
2025-02-01T11:20:25Z [      main] knst getAverageAnonymizedRounds() 35.45ms
2025-02-01T11:20:28Z [      main] knst getBalances() 142.88ms
2025-02-01T11:20:28Z [      main] knst getBalances() 137.15ms
2025-02-01T11:20:29Z [      main] knst getBalances() 136.33ms
2025-02-01T11:20:29Z [      main] knst getBalances() 137.43ms
2025-02-01T11:20:29Z [      main] knst getAnonymizableBalance() 172.26ms
2025-02-01T11:20:29Z [      main] knst getBalances() 140.42ms
2025-02-01T11:20:29Z [      main] knst getNormalizedAnonymizedBalance() 25.78ms
2025-02-01T11:20:29Z [      main] knst getAverageAnonymizedRounds() 34.88ms
2025-02-01T11:20:29Z [      main] knst getBalances() 144.52ms
2025-02-01T11:20:29Z [      main] knst getAnonymizableBalance() 170.65ms
2025-02-01T11:20:30Z [      main] knst getBalances() 141.37ms
2025-02-01T11:20:30Z [      main] knst getNormalizedAnonymizedBalance() 26.05ms
2025-02-01T11:20:30Z [      main] knst getAverageAnonymizedRounds() 36.16ms
2025-02-01T11:20:30Z [      main] knst getBalances() 143.27ms

3404  +0  +544.79  +173.77  +139.59  +25.87  +35.35  +171.82  +141.07  +25.77  +35.45  +142.88  +137.15  +136.33  +137.43  +172.26  +140.42  +25.78  +34.88  +144.52  +170.65  +141.37  +26.05  +36.16  +143.27
=6286.63
```

Sum-up improvement is slightly higher then real improvement of call `loadwallet` (1 second as sum of function vs 0.5 second as total RPC time) due to fact that calculations distributed between 2 threads: `main` and `qt-rpcconsole`.


## Breaking Changes
N/A

## Checklist:
- [x] I have performed a self-review of my own code
- [x] I have commented my code, particularly in hard-to-understand areas
- [ ] I have added or updated relevant unit/integration/functional/e2e tests
- [ ] I have made corresponding changes to the documentation
- [x] I have assigned this pull request to a milestone