@knst 2025-02-04
> pls see https://github.com/dashpay/dash/commit/66089d0e65023b14489412a1bd75742c7212db97 and https://github.com/dashpay/dash/commit/c32ca0bf33b048001f7186f984ca76644b463dd2

Good catch, there's a bug has been added by me when I rebased PR to make it review-ready. `GetBalanceAnonymized` is used now inside `getAvailableBalance`.